Integrators

The integrators used in PoMiN

RK4 Integrator

pomin.Jsympl — Function

Jsympl( Zarg::RealVec )

Symplectic operator $\hat{J}$. Takes vector $\vec{z}$ and returns $\hat{J}\vec{z}$ where $\hat{J} = \begin{bmatrix} 0 & I \ -I & 0 \end{bmatrix}$

pomin.hrkintegrator — Function

hrkintegrator( z0::RealVec, dH::Function , δ::Real 
                , tadap::Function , tspan::Tuple{Real,Real} , maxit::Real )

4th order Runge-Kutta integrator

Arguments

z0::RealVec: Initial values phase space vector $\{ \vec{q}, \vec{p} \}$

dH::Function: Gradient of the Hamiltonian with respect to the phase space vector $\vec{z}$. Takes a single parameter $\vec{zi}$ of the values of the phase space variables at the current time and returns a the gradient of H which is a vector with the same dimensionality as $\vec{zi}$

δ::Real: Time step to be used by the integrator

tadap::Function: Adaptive time-stepping function. Takes a single parameter $\vec{zi}$ of the values of the phase space variables at the current time and returns the time step

tspan::Tuple{Real,Real}: Tuple containing the start time and end time for the integration

maxit::Real: Maximum number of iterations. When this number of iterations is exceeded, integration stops.

Tao Symplectic Integrator

pomin.hsintegrator — Function

hsintegrator( z0::RealVec, dH::Function , δ::Real , ω::Real , tspan::Tuple{Real,Real} , maxit::Real )

Symplectic integrator of Tao

Arguments

z0::RealVec: Initial values of phase space vector $\{ \vec{q}, \vec{p} \}$

dH::Function: Gradient of the Hamiltonian with respect to the phase space vector $\vec{z}$. Takes a single parameter $\vec{zi}$ of the values of the phase space variables at the current time and returns a the gradient of H which is a vector with the same dimensionality as $\vec{zi}$

δ::Real: Time step to be used by the integrator

ω::Real: Parameter needed for Tao's method of integration. See arXiv:1609.02212

tspan::Tuple{Real,Real}: Tuple containing the start time and end time for the integration

maxit::Real: Maximum number of iterations. When this number of iterations is exceeded, integration stops.